Can add journeys
This is not perfect yet. It is not possible to add journeys from csv if the insertion
of any one row does not succeed.

Maybe create a seperate endpoint for inserting csvs or maybe add a migration where the
insertion of csvs is done. Could use git LFS to store the csvs in the repository.

Could also create something in the frontend to import csvs. The validation logic
is needed there anyway.

There might also be a problem with the generated id. It might clash because
insertions are now just done with their ids. Could prevent this by not allowing
insertion of ids. Could add a seperate table or column to store the id and just generate
a new one for every insertion. Could also just set the sequence high enough by setting it to
max(inserted_id+1, current_value).